evaluators/text/EditDistanceEvaluator/requirements.txt
jina>=0.7
python-Levenshtein-wheels
